# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/services": {
      "post": {
        "operationId": "create",
        "summary": "Create new service",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Service created successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all services",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return all services"
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/services/provider": {
      "get": {
        "operationId": "findProviderServices",
        "summary": "Get provider services",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Return provider services"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/category/{categoryId}": {
      "get": {
        "operationId": "findByCategory",
        "summary": "Get services by category",
        "parameters": [
          {
            "name": "categoryId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return services by category"
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/services/nearby": {
      "get": {
        "operationId": "findNearby",
        "summary": "Get nearby services",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "distance",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return nearby services"
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
    "/services/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get service by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return service details"
          }
        },
        "tags": [
          "Services"
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update service",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Service updated successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete service",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service deleted successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update service status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service status updated successfully"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/services/{id}/featured": {
      "put": {
        "operationId": "toggleFeatured",
        "summary": "Toggle service featured status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Service featured status updated"
          }
        },
        "tags": [
          "Services"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/categories/{id}/featured": {
      "put": {
        "operationId": "toggleFeatured",
        "summary": "Toggle category featured status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category featured status updated"
          }
        },
        "tags": [
          "Categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/payments/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update payment status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaymentStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment status updated"
          }
        },
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/reviews/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update review status (Admin)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Review status updated successfully"
          }
        },
        "tags": [
          "Reviews"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/settings": {
      "post": {
        "operationId": "create",
        "summary": "Create new setting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSettingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Setting created successfully"
          }
        },
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all settings",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return all settings"
          }
        },
        "tags": [
          "Settings"
        ]
      }
    },
...
    "/settings/values": {
      "get": {
        "operationId": "getValues",
        "summary": "Get multiple setting values",
        "parameters": [
          {
            "name": "keys",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return setting values"
          }
        },
        "tags": [
          "Settings"
        ]
      }
    },
    "/settings/{key}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get setting by key",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return setting details"
          }
        },
        "tags": [
          "Settings"
        ]
      },
      "put": {
        "operationId": "update",
        "summary": "Update setting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Setting updated successfully"
          }
        },
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete setting",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setting deleted successfully"
          }
        },
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/settings/bulk": {
      "put": {
        "operationId": "bulkUpdate",
        "summary": "Bulk update settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Settings updated successfully"
          }
        },
        "tags": [
          "Settings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/notifications/{id}/status": {
      "put": {
        "operationId": "updateStatus",
        "summary": "Update notification status",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notification status updated"
          }
        },
        "tags": [
          "Notifications"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "HizmetBul API",
    "description": "HizmetBul Servis Platformu API Dokümantasyonu",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LocationPropertiesDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "example": "İstiklal Caddesi No: 50, Beyoğlu"
          },
          "city": {
            "type": "string",
            "example": "İstanbul"
          },
          "state": {
            "type": "string",
            "example": "İstanbul"
          }
        },
        "required": [
          "address",
          "city",
          "state"
        ]
      },
      "LocationDto": {
        "type": "object",
        "properties": {
          "lng": {
            "type": "number",
            "example": 28.9784
          },
          "lat": {
            "type": "number",
            "example": 41.0082
          },
          "properties": {
            "$ref": "#/components/schemas/LocationPropertiesDto"
          }
        },
        "required": [
          "lng",
          "lat",
          "properties"
        ]
      },
      "WorkingHoursDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "example": "09:00"
          },
          "end": {
            "type": "string",
            "example": "18:00"
          }
        },
        "required": [
          "start",
          "end"
        ]
      },
      "AvailabilityDto": {
        "type": "object",
        "properties": {
          "workingHours": {
            "example": {
              "start": "09:00",
              "end": "18:00"
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/WorkingHoursDto"
              }
            ]
          },
          "workingDays": {
            "example": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "exceptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "workingHours",
          "workingDays"
        ]
      },
      "SettingsDto": {
        "type": "object",
        "properties": {
          "instantBooking": {
            "type": "boolean",
            "example": true
          },
          "minNotice": {
            "type": "number",
            "example": 2
          },
          "maxBookingsPerDay": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "instantBooking",
          "minNotice",
          "maxBookingsPerDay"
        ]
      },
      "CreateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Profesyonel Ev Temizliği"
          },
          "description": {
            "type": "string",
            "example": "Detaylı ev temizlik hizmeti..."
          },
          "categoryId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "price": {
            "type": "number",
            "example": 250
          },
          "currency": {
            "type": "string",
            "example": "TRY"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "pending",
              "rejected"
            ]
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityDto"
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsDto"
          }
        },
        "required": [
          "title",
          "description",
          "categoryId",
          "price",
          "location",
          "availability",
          "settings"
        ]
      },
      "UpdateServiceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Profesyonel Ev Temizliği"
          },
          "description": {
            "type": "string",
            "example": "Detaylı ev temizlik hizmeti..."
          },
          "categoryId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "price": {
            "type": "number",
            "example": 250
          },
          "currency": {
            "type": "string",
            "example": "TRY"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "pending",
              "rejected"
            ]
          },
          "availability": {
            "$ref": "#/components/schemas/AvailabilityDto"
          },
          "settings": {
            "$ref": "#/components/schemas/SettingsDto"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "UpdatePaymentStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              "refunded"
            ]
          },
          "transactionId": {
            "type": "string",
            "example": "TR123456789"
          },
          "error": {
            "type": "string",
            "example": "Payment failed: Insufficient funds"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "UpdateReviewStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "ValidationDto": {
        "type": "object",
        "properties": {
          "required": {
            "type": "boolean"
          },
          "min": {
            "type": "number"
          },
          "max": {
            "type": "number"
          },
          "pattern": {
            "type": "string"
          }
        }
      },
      "MetadataDto": {
        "type": "object",
        "properties": {
          "inputType": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "validation": {
            "$ref": "#/components/schemas/ValidationDto"
          },
          "group": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        }
      },
      "CreateSettingDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "example": "site.title"
          },
          "value": {
            "type": "string",
            "example": "HizmetBul"
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "mail",
              "payment",
              "notification",
              "social",
              "integration"
            ]
          },
          "description": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "default": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        },
        "required": [
          "key",
          "value",
          "type"
        ]
      },
      "UpdateSettingDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "example": "HizmetBul"
          },
          "type": {
            "type": "string",
            "enum": [
              "general",
              "mail",
              "payment",
              "notification",
              "social",
              "integration"
            ]
          },
          "description": {
            "type": "string"
          },
          "isPublic": {
            "type": "boolean",
            "default": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataDto"
          }
        }
      },
      "UpdateNotificationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "unread",
              "read",
              "archived"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}